wip(batch 1/4): E-W1/M-1119 -- lift ternary conversion-utility ceiling i64->i128 + W-1 sweep
mycelium-core::ternary and mycelium-std-ternary::arithmetic's max_magnitude/int_to_trits/
trits_to_int widen from i64 to i128 (KC-3, no unsafe): the real pre-widening ceiling was
m<=39, not the documented m<=40 (max_magnitude's naive 3^m computation overflowed i64 one
trit earlier than claimed -- confirmed by direct computation, matches
mycelium-mlir::swap_codegen::MAX_TERNARY_WIDTH_I64's independent m<=39 figure). A second,
more serious pre-existing gap: the old i64 trits_to_int Horner fold could transiently
overflow mid-decode for a genuinely 41-trit-significant value (e.g. i64::MIN at m=41) even
though the final result fit i64 -- a live G2 gap mycelium-cert::ternary_to_binary would hit
unconditionally once a Ternary{41} value entered the system. Both are fixed by the i128
widening (safe through m<=80; m>=81 is the new explicit None).

Downstream call sites updated to agree (mycelium-cert::legal_pair/binary_to_ternary/
ternary_to_binary, mycelium-mlir::swap_codegen::legal_pair, plus test call sites across
mycelium-cert/mycelium-mlir/mycelium-interp/mycelium-std-swap/mycelium-std-conformance/xtask).
Inline tests in the touched logic files (ternary/mod.rs, ternary/big_ternary.rs,
mycelium-std-ternary/src/arithmetic.rs) extracted to src/tests/ per the house test-layout
rule (as-touched, M-797), with new m=41 exact round-trip tests (the DoD's None->Some flip
witness) plus a new ceiling test at the widened m=80/81 boundary.

W-1 canon sweep (docs/spec/swaps/binary-ternary.md + docs/spec/stdlib/swap.md 2026-07-18
amendments): lib/std/swap.myc gains bin32_to_tern21/tern21_to_bin32 (available since M-111,
no gap) and bin64_to_tern41/tern41_to_bin64 (unblocked by this fix) via the native
swap(..., to: ..., policy: ...) construct, verified myc-check-clean and exercised end to end
(L1-eval/L0-interp/AOT + live Rust oracle) over a curated Binary{64} edge corpus in a new
mycelium-std-conformance test. matrix_len moved Binary{8}->Binary{64} per swap.md's length/
count canon (with its own conformance test updated); bytes_len's matching Binary{32}->
Binary{64} move is a kernel-primitive-registry change, deliberately ledgered as a residual
rather than forced (out of this narrower ceiling-lift's scope). docs/spec/stdlib/ternary.md
gets an append-only amendment recording the breaking i64->i128 Rust-signature change (ADR-045
S2 obligation on the DN-66-frozen surface); docs/lib-index/ regenerated in-session.

Advances M-1119 (E-W1) DoD 1-3 and the course-correction W-D item 3 (W-1 Section 2.1 sweep).
Verified: cargo test/clippy/fmt green on every touched crate (mycelium-core,
mycelium-std-ternary, mycelium-cert, mycelium-mlir, mycelium-interp, mycelium-std-swap,
mycelium-std-conformance, mycelium-l1, xtask); myc-check clean on lib/std.

FLAG for the integrating parent: M-1119 status flip todo->done in tools/github/issues.yaml
(landed basis: this commit); CHANGELOG.md entry (breaking mycelium-std-ternary signature
change, noted in ternary.md's own amendment); docs/api-index/ regen (mycelium-cert's public
fn signatures changed).

+++ xtask/src/kc4.rs
@@ -78,7 +78,7 @@ fn byte_of(v: i64) -> Value {
 fn tern_of(v: i64, trits: u32) -> Value {
     Value::new(
         Repr::Ternary { trits },
-        Payload::Trits(ternary::int_to_trits(v, trits).unwrap()),
+        Payload::Trits(ternary::int_to_trits(i128::from(v), trits).unwrap()),
         Meta::exact(Provenance::Root),
     )
     .unwrap()
